refactor(core): added baseline
