+++ .bumpversion.cfg
@@ -1,0 +1,6 @@
+[bumpversion]
+current_version = 1.0.1
+commit = True
+tag = False
+
+[bumpversion:file:pyproject.toml]
