+++ .gitattributes
@@ -1,0 +1,5 @@
+*.yml text eol=lf
+*.yaml text eol=lf
+*.py text eol=lf
+*.md text eol=lf
+*.txt text eol=lf
