+++ .github/FUNDING.yml
@@ -1,0 +1,17 @@
+---
+# .github/FUNDING.yml
+
+# These are supported funding model platforms
+github: [JuanVilla424]
+#patreon: # Replace with a single Patreon username
+#open_collective: # Replace with a single Open Collective username
+ko_fi: na0nh
+#tidelift: # Replace with a single Tidelift platform-name/package-name e.g., npm/babel
+#community_bridge: # Replace with a single Community Bridge project-name e.g., cloud-foundry
+#liberapay: # Replace with a single Liberapay username
+#issuehunt: # Replace with a single IssueHunt username
+#lfx_crowdfunding: # Replace with a single LFX Crowdfunding project-name e.g., cloud-foundry
+#polar: # Replace with a single Polar username
+#buy_me_a_coffee: # Replace with a single Buy Me a Coffee username
+#thanks_dev: # Replace with a single thanks.dev username
+#custom: # Replace with up to 4 custom sponsorship URLs e.g., ['link1', 'link2']
