+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,11 @@
+---
+# .github/ISSUE_TEMPLATE/config.yml
+
+blank_issues_enabled: false
+contact_links:
+  - name: Langding Community Support
+    url: https://github.com/JuanVilla424/langding/discussions
+    about: Please ask and answer questions here.
+  - name: Langding Security Reports
+    url: https://juanvilla424.github.io/langding/SECURITY
+    about: Please report security vulnerabilities here.
