+++ .github/dependabot.yml
@@ -1,0 +1,20 @@
+# .github/dependabot.yml
+
+version: 2
+updates:
+  - package-ecosystem: "pip"
+    directory: "."
+    schedule:
+      interval: "weekly"
+    commit-message:
+      prefix: "deps"
+    open-pull-requests-limit: 5
+    target-branch: "dev"
+  - package-ecosystem: "pip"
+    directory: "backend"
+    schedule:
+      interval: "weekly"
+    commit-message:
+      prefix: "deps"
+    open-pull-requests-limit: 5
+    target-branch: "dev"
