+++ bjorn-detector.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+
+from src import main
+
+main.main()
