+++ pyproject.toml
@@ -1,0 +1,74 @@
+[tool.poetry]
+name = "bjorn-detector"
+version = "1.0.1"
+description = "One-Click SSH access to your bjorn device in a radar style."
+# authors =
+license = "MIT"
+readme = "README.md"
+package-mode = false
+
+[tool.poetry.dependencies]
+python = "^3.9"
+setuptools = "^75.2.0"
+idna="^3.0"
+certifi="^2024.8.30"
+bump2version="^1.0.0"
+PyQt6="^6.7.0"
+
+[tool.poetry.group.dev.dependencies]
+pre-commit = "^4.0.0"
+pylint="^3.3.0"
+isort = "^5.12.0"
+toml="^0.10.0"
+black="^24.3.0"
+pytest = "^8.3.1"
+httpx = "^0.27.2"
+pytest-cov = "^6.0.0"
+alembic = "^1.11.1"
+yamllint = "^1.35.0"
+coverage = "^7.2.5"
+
+[tool.poetry.extras]
+migrations = ["alembic"]
+testing = ["pytest", "httpx"]
+
+[tool.black]
+line-length = 100
+target-version = ['py39']
+include = '\.pyi?$'
+exclude = '''
+/(
+    \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | venv
+  | _build
+  | buck-out
+  | build
+  | dist
+  | node_modules
+)/
+'''
+
+[tool.pytest.ini_options]
+minversion = "6.0"
+addopts = "--cov=app --cov-report=xml:coverage.xml --cov-report=term"
+testpaths = ["tests"]
+
+[tool.isort]
+profile = "black"
+line_length = 100
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+
+[tool.pylint]
+rcfile = ".pylintrc"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
