+++ requirements.txt
@@ -1,1 +1,1 @@
-PyQt6
+PyQt6>=6.7.0
